// packages/commercial-lytenyte-pro/src/components/external/floating-ui.d.ts
declare type ClientRectObject = Prettify$1<Rect & SideObject>;
declare type Coords = { [key in Axis]: number };
declare type Dimensions = { [key in Length]: number };
declare interface ElementRects {
  reference: Rect;
  floating: Rect;
}
declare type Length = "width" | "height";
declare type Padding = number | Prettify$1<Partial<SideObject>>;
declare type Placement = Prettify$1<Side | AlignedPlacement>;
declare type Prettify$1<T> = { [K in keyof T]: T[K] } & {};
declare type Rect = Prettify$1<Coords & Dimensions>;
declare type Side = "top" | "right" | "bottom" | "left";
declare type SideObject = { [key in Side]: number };
// ...
declare interface VirtualElement {
  getBoundingClientRect(): ClientRectObject;
  getClientRects?(): Array<ClientRectObject> | DOMRectList;
  contextElement?: Element;
}